# ChatGPT CodeReview This GitHub Action uses the ChatGPT API to review code in Pull Requests. It will trigger on pull request events when opened or synchronized.
## hello_world_job
This job will run on an Ubuntu-latest runner and is named ChatGPT explain code. 

## Steps
The step uses the ChatGPT Github Actions from the cirolini repository. It will use the OpenAI API key and Github Token from the repository secrets and the Github Pull Request ID from the Github event. It is also possible to specify the OpenAI engine to use and the sampling temperature and maximum number of tokens to generate in OpenAI completion.