# ChatGPT CodeReview This GitHub Action uses [ChatGPT](https://chatgpt.com/) to review code in a pull request. It is triggered on pull request events when opened or synchronized.
The job is run on an Ubuntu-latest runner and is named `ChatGPT explain code`.

The step uses the ChatGPT Github Actions from the cirolini repository. The OpenAI API key and Github Token are retrieved from the repository secrets, while the Github Pull Request ID is taken from the Github event.

The OpenAI engine, sampling temperature, and maximum number of tokens to generate in OpenAI completion are all optional and have default values.