# ChatGPT CodeReview This GitHub Action runs ChatGPT to explain code in Pull Requests.
It is triggered on pull request events (opened, synchronized) and runs on an Ubuntu-latest runner. 

The Action uses the ChatGPT Github Actions from the cirolini repository and requires the OpenAI API key and Github Token from repository secrets. It also takes in the Github Pull Request ID from the Github event. 

Optional parameters include specifying the OpenAI engine, sampling temperature and maximum number of tokens to generate in OpenAI completion. The defaults are "text-davinci-002", 0.5 and 2048 respectively.