# ChatGPT CodeReview
This GitHub Action is used to trigger a ChatGPT explain code job when a pull request is opened or synchronized. This job will use the OpenAI API key and GitHub token from the repository secrets, as well as the Github Pull Request ID from the Github event. Additionally, the OpenAI engine and sampling temperature can be specified (default is "text-davinci-002" and 0.5 respectively). The maximum number of tokens to generate in OpenAI completion can also be specified (default is 2048). Finally, the pull request base ref is also provided.