# ChatGPT CodeReview This GitHub Action runs ChatGPT on Pull Request events to explain the code.
It uses the OpenAI API key and Github Token from repository secrets and the Github Pull Request ID from the Github event. 

Optionally, it can specify the OpenAI engine and sampling temperature for OpenAI completion. The default values are "text-davinci-002" and 0.5 respectively. The maximum number of tokens to generate in OpenAI completion is 2048 by default.