# ChatGPT CodeReview This GitHub Action runs a code review using ChatGPT. It is triggered when a pull request is opened or synchronized and runs on an Ubuntu-latest runner.
# Job
This job is called "ChatGPT explain code" and uses the ChatGPT Github Actions from cirolini repository.

# Steps
This step is called "ChatGPT explain code" and uses the ChatGPT Github Actions from cirolini repository. It uses the OpenAI API key from the repository secrets, the Github Token from the repository secrets, the Github Pull Request ID from the Github event, the OpenAI engine to use (default is "text-davinci-002"), the sampling temperature for OpenAI (default is 0.5) and the maximum number of tokens to generate in OpenAI completion (default is 2048).